% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numeric.R
\name{format_numeric}
\alias{format_numeric}
\title{helper function to format output to csvs}
\usage{
format_numeric(x, ...)
}
\arguments{
\item{x}{dataframe to format}
}
\description{
helper function to format output to csvs
}
\details{
base::format does not appear to produce a data.frame so it can't be used with write_tsv. I typically use this with sci=FALSE to turn off scientific notation but it can also control significant digits, decimal places etc. see `format` for options.
}
