% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_station_inventory.R
\name{get_inv}
\alias{get_inv}
\title{setup and clean station inventory}
\usage{
get_inv(network = "snotel")
}
\arguments{
\item{network}{which station network are you using. default 'snotel'. currently only option}
}
\value{
tibble with (at least) column names Site_ID, site_name, Latitude, Longitude.
}
\description{
setup and clean station inventory
}
\details{
There is a snotel inventory included with the package from Jan 2017. It can be updated from https://wcc.sc.egov.usda.gov/nwcc/inventory or theoretically replaced with that of a different network
}
