% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modscag_data.R
\name{get_modscag_data}
\alias{get_modscag_data}
\title{Function to get MODSCAG data.}
\usage{
get_modscag_data(doy = NULL, yr = NULL, type = NULL, imagepath = NULL,
  reso, xmin, xmax, ymin, ymax)
}
\arguments{
\item{doy}{day of year}

\item{yr}{year}

\item{type}{'NRT' or 'historic'. near realtime download from snowdav or archived data}

\item{imagepath}{the directory where the mosaiced fsca .tif files exist}

\item{reso}{resolution of fsca}

\item{xmin}{minimum x coordinate of domain}

\item{xmax}{maximum x coordinate of domain}

\item{ymin}{minimum y coordinate of domain}

\item{ymax}{maximum y coordinate of domain}
}
\description{
The function will check for existing processed fsca images in the imagepath. If it doesn't exist, it will check for raw, unprocessed files and mosaic them. If no file exist it will attempt to download them from snowserver and mosaic them.
}
\details{
called for its side effect of producing a mosaicked fsca image of the domain
}
