% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_rcn_nc.R
\name{get_rcn_nc}
\alias{get_rcn_nc}
\title{function to extract the 1st and 15th of each month from a single year of reconstruction}
\usage{
get_rcn_nc(ryr)
}
\arguments{
\item{ryr}{the year of reconstruction}
}
\value{
rasterstack of reconstructed swe for the 1st and 15th of each year present in recondata*.nc files
}
\description{
function to extract the 1st and 15th of each month from a single year of reconstruction
}
\details{
It's assumed that the .nc file has dimensions Lat, Long, time. The variable is "swe" in meters. Values of the time dimension are stored YYYYMMDD.
}
