% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_stationswe_data.R
\name{get_stationswe_data}
\alias{get_stationswe_data}
\title{download and combine station data}
\usage{
get_stationswe_data(yr = 2017, station_locs, network = "snotel")
}
\arguments{
\item{yr}{yr of the simulation date}

\item{station_locs}{tbl_df of site metadata}

\item{network}{default snotel}
}
\value{
tibble with station data from nrcs
}
\description{
download and combine station data
}
