% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gnet_phvfsca.R
\name{gnet_phvfsca}
\alias{gnet_phvfsca}
\title{Fit the GLM-Net model to the data}
\usage{
gnet_phvfsca(dF, cl = NULL)
}
\arguments{
\item{dF}{dataframe that includes x and y variables for regression}

\item{cl}{a cluster object for parallel execution}
}
\description{
Fit the GLM-Net model to the data
}
