% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_dates.R
\name{import_dates}
\alias{import_dates}
\title{reads dates from a csv file}
\usage{
import_dates(fn)
}
\arguments{
\item{fn}{filename of csv file with dates}
}
\value{
a tibble of dates with new columns for yr, doy, day
}
\description{
reads dates from a csv file
}
\details{
the csv should have exactly 1 column with a column heading. the date format must be year-numeric_month-day
}
