% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_watermask.R
\name{load_watermask}
\alias{load_watermask}
\title{Load the watermask}
\usage{
load_watermask(runname)
}
\arguments{
\item{runname}{RUNNAME specified in the run file}
}
\value{
rasterImage of watermask
}
\description{
This function will load the watermask for the given RUNNAME and compare extents vs given extents.
}
