% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{myr2}
\alias{myr2}
\title{r-square}
\usage{
myr2(model, data)
}
\arguments{
\item{model}{model object}

\item{data}{dataframe to predict with modelobj}
}
\description{
r-square
}
\details{
there are two rsquare functions because I clipped them from different pieces of code. They take different types of inputs but perform the same calculation.
}
