% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-num2ucoRaster.R
\name{num2ucoRaster}
\alias{num2ucoRaster}
\title{num2ucoRaster}
\usage{
num2ucoRaster(v, nr = 2580)
}
\arguments{
\item{v}{numeric vector}

\item{nr}{number of rows, default 2580}
}
\value{
raster object for UCO domain
}
\description{
function to convert vector to raster
}
