% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pickAlpha.R
\name{pickAlpha}
\alias{pickAlpha}
\title{select alpha for glmnet by crossvalidation}
\usage{
pickAlpha(dF, myformula, nfolds, cl = cl)
}
\arguments{
\item{dF}{dataframe with x and y variables for regression}

\item{myformula}{regression formula}

\item{nfolds}{number of folds for k-fold cross validation}

\item{cl}{cluster object for parallel processing}
}
\description{
select alpha for glmnet by crossvalidation
}
\details{
this is a helper function for gnet_phvfsca
}
