% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{r2}
\alias{r2}
\title{r-square}
\usage{
r2(yobs, yhat)
}
\arguments{
\item{yobs}{observed}

\item{yhat}{predicted}
}
\description{
r-square
}
\details{
there are two rsquare functions because I clipped them from different pieces of code. They take different types of inputs but perform the same calculation.
}
