% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-readgrads.R
\name{readgrads}
\alias{readgrads}
\title{read grads binary cube}
\usage{
readgrads(t, fid = fid, nc = 1950, nr = 2580)
}
\arguments{
\item{t}{timestep}

\item{fid}{file connection}

\item{nc}{number of columns, default 1950}

\item{nr}{number of rows, default 2580}
}
\description{
function to read grads formatted binary cube to raster
}
