% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_modeldata.R
\name{setup_modeldata}
\alias{setup_modeldata}
\title{setup modeldata based on fsca or rcn as predictor}
\usage{
setup_modeldata(snoteltoday.sp, snoteltoday, phvsnotel, myformula, simfsca,
  snow_var, PATH_RCNDOWNLOAD)
}
\arguments{
\item{snoteltoday.sp}{spatialpoints data frame of the snotel data for the simulation date}

\item{snoteltoday}{regular dataframe  of the snotel data for the simulation date}

\item{phvsnotel}{dataframe containing the physiographic values for the pixels that contain snotel stations}

\item{myformula}{formula for regression}

\item{simfsca}{raster. modscag image of fsca for the simulation date}

\item{snow_var}{character string identifying the snow related regression variable}

\item{PATH_RCNDOWNLOAD}{path for the location of recondata*.nc files used with snow_var is 'rcn'}
}
\description{
setup modeldata based on fsca or rcn as predictor
}
