% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-unscale_vec.R
\name{unscale_vec}
\alias{unscale_vec}
\title{unscale vector}
\usage{
unscale_vec(vec, scale_values = NULL)
}
\arguments{
\item{vec}{numeric vector to unscale}

\item{scale_values}{vector with scale attributes to use}
}
\description{
function to unscale column of dataframe
}
