% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lib-water_yr.R
\name{water_yr}
\alias{water_yr}
\title{wateryear}
\usage{
water_yr(dates, start_month = 10)
}
\arguments{
\item{dates}{date to convert}

\item{start_month}{defaults to 10}
}
\value{
wateryear of the date
}
\description{
function to calculate wateryear from date column
}
