% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modscag_data.R
\name{get_modscag_data}
\alias{get_modscag_data}
\title{Function to get MODSCAG data
Either download the near realtime data from snowdav system or use archived historic data}
\usage{
get_modscag_data(doy = NULL, yr = NULL, type = NULL, imagepath = NULL)
}
\arguments{
\item{doy}{day of year}

\item{yr}{year}

\item{source}{'NRT' or 'historic'. near realtime download from snowdav or archived data}
}
\description{
Function to get MODSCAG data
Either download the near realtime data from snowdav system or use archived historic data
}
\details{
called for its side effect of producing a mosaicked fsca image of the domain
}
