% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_modscag_data.R
\name{get_modscag_data}
\alias{get_modscag_data}
\title{Function to get MODSCAG data.}
\usage{
get_modscag_data(doy, yr, type, imagepath, modscagfn = "snow_fraction",
  fveg_correction = TRUE, reso, xmin, xmax, ymin, ymax)
}
\arguments{
\item{doy}{day of year}

\item{yr}{year}

\item{type}{'NRT' or 'historic'. near realtime download from snowdav or archived data}

\item{imagepath}{the directory where the mosaiced fsca .tif files exist}

\item{modscagfn}{'snow_fraction' or 'snow_fraction_canadj'; portion of the filename before the extension to identify canopy corrected or not corrected modscag;}

\item{fveg_correction}{logical, default TRUE. should fsca be corrected by modscag fveg?}

\item{reso}{resolution of fsca}

\item{xmin}{minimum x coordinate of domain}

\item{xmax}{maximum x coordinate of domain}

\item{ymin}{minimum y coordinate of domain}

\item{ymax}{maximum y coordinate of domain}
}
\value{
rasterImage of fsca for the simulation date
}
\description{
The function will check for existing processed fsca images in the imagepath. If it doesn't exist, it will check for raw, unprocessed files and mosaic them. If no file exist it will attempt to download them from snowserver and mosaic them.
}
\details{
called for its side effect of producing a mosaicked fsca image of the domain
}
