% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_functions.R
\name{mypctmae}
\alias{mypctmae}
\title{error functions}
\usage{
mypctmae(model, data)
}
\arguments{
\item{model}{model object}

\item{data}{dataframe to predict with modelobj}
}
\description{
error functions
}
